+++ src/@classes/abstract/IsometricShape/types.ts
@@ -1,0 +1,6 @@
+import { IsometricPlaneView } from '@types';
+import { GraphicProps } from '@classes/abstract/Graphic';
+
+export interface IsometricShapeProps extends GraphicProps {
+    planeView: IsometricPlaneView;
+}