+++ src/@classes/public/IsometricRectangle/types.ts
@@ -1,0 +1,6 @@
+import { IsometricShapeProps } from '@classes/abstract/IsometricShape';
+
+export interface IsometricRectangleProps extends IsometricShapeProps {
+    height: number;
+    width: number;
+}