+++ src/@types/index.ts
@@ -28,3 +28,31 @@ export interface Listener {
     fn: VoidFunction;
     fnBind: VoidFunction;
 }
+
+export enum Command {
+    move = 'move',
+    line = 'line',
+    curve = 'curve'
+}
+
+export interface LinePoint {
+    command: Exclude<Command, Command.curve>;
+    control?: never;
+    point: Point;
+}
+
+export interface CurvePoint {
+    command: Exclude<Command, Command.move | Command.line>;
+    control: Point;
+    point: Point;
+}
+
+export type CommandPoint = LinePoint | CurvePoint;
+
+export enum PlaneView {
+    FRONT = 'FRONT',
+    SIDE = 'SIDE',
+    TOP = 'TOP'
+}
+
+export type IsometricPlaneView = keyof typeof PlaneView;